requirements.txt
requests
msal
schedule
openai
azure-identity
